Fixes for build with mingw (no warnings in 32 Bit build, even with -wall). SMD bugfix. Makefile is up to date now. Added pack-attribute for aiVector3D and aiVector2D.
git-svn-id: https://assimp.svn.sourceforge.net/svnroot/assimp/trunk@132 67173fc5-114c-0410-ac8e-9d2fd5bffc1f

+++ code/ImproveCacheLocality.cpp
@@ -271,7 +271,7 @@ void ImproveCacheLocalityProcess::ProcessMesh( aiMesh* pMesh, unsigned int meshN
 					const unsigned int dp = *p;
 
 					// the current vertex won't have any free triangles after this step
-					if (ivdx != dp)
+					if (ivdx != (int)dp)
 					{
 						// append the vertex to the dead-end stack
 						sDeadEndVStack.push(dp);
